Fix issue with django liveserver + react liveserver
It seems django-cra-helper doesn't have the right files in the
`bundle_js` array to get live-reloading working (at least for
react-scripts@4.x). Workaround by hardcoding the tag for the missing
script.

This is worth opening an issue in django-cra-helper, seems like either
an easy fix for them or a stupid mistake on my part.